% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sora.R
\name{is_shared}
\alias{is_shared}
\title{Test if an Object is Shared}
\usage{
is_shared(x)
}
\arguments{
\item{x}{an R object.}
}
\value{
\code{TRUE} or \code{FALSE}.
}
\description{
Returns \code{TRUE} if \code{x} is an ALTREP object backed by shared memory
(created by \code{\link[=sora]{sora()}} or \code{\link[=map_shared]{map_shared()}}), \code{FALSE} otherwise.
}
\examples{
x <- sora(rnorm(100))
is_shared(x)
is_shared(rnorm(100))

}
